[Syntax] move parsing on expressions
